+++ src/components/BlogList/BlogListItem/BlogListItem.tsx
@@ -7,6 +7,7 @@ import Box from "../../Box";
 import CMSImage from "../../CMSImage";
 import Flex from "../../Flex";
 import LineClamp from "../../LineClamp";
+import BoxBorders from "../../SpriteSheet/BrushSvgs/BoxBorders";
 import { P, Heading, HeadingTag } from "../../Typography";
 
 const ActionLink = styled.a`
@@ -52,25 +53,39 @@ const BlogListItem: FC<BlogListItemProps> = (props) => {
     mainImage,
   } = props;
 
-  const blogDate = new Date(date);
+  const blogDate = new Date(date).toLocaleDateString("en-GB", {
+    day: "numeric",
+    month: "long",
+    year: "numeric",
+  });
 
   return (
-    <Flex $position={"relative"} $mt={24} $flexDirection={["column", "row"]}>
+    <Flex
+      $position={"relative"}
+      $flexDirection={["column", "row"]}
+      $alignItems={["initial", "center"]}
+      $minHeight={200}
+    >
       {withImage && mainImage && (
-        <Box $position={"relative"} $minWidth={240}>
-          <AspectRatio ratio={"3:2"}>
-            <CMSImage
-              layout="fill"
-              objectFit="cover"
-              objectPosition="center center"
-              image={mainImage}
-              // @TODO: Replace with valid alt text when supported by the CMS
-              alt=""
-            />
-          </AspectRatio>
+        <Box $position={"relative"} $minWidth={240} $mr={[0, 32]} $mb={[32, 0]}>
+          <BoxBorders />
+          <Box $ma={1}>
+            <AspectRatio ratio={"3:2"}>
+              <CMSImage
+                layout="fill"
+                objectFit="cover"
+                objectPosition="center center"
+                image={mainImage}
+                sizes="(min-width: 750px) 256px, 100vw"
+                // @TODO: Replace with valid alt text when supported by the CMS
+                alt=""
+              />
+            </AspectRatio>
+          </Box>
         </Box>
       )}
-      <Flex $flexDirection="column" $alignItems="flex-start" $ml={[0, 32]}>
+
+      <Flex $flexDirection="column" $alignItems="flex-start">
         <P
           $fontSize={16}
           $lineHeight={"20px"}
@@ -81,18 +96,14 @@ const BlogListItem: FC<BlogListItemProps> = (props) => {
           {category.title}
         </P>
         <P $fontSize={14} $lineHeight={"20px"} $mt={16}>
-          {blogDate.toLocaleDateString("en-GB", {
-            day: "numeric",
-            month: "long",
-            year: "numeric",
-          })}
+          {blogDate}
         </P>
-        <Heading tag={titleTag} $fontSize={20} $mt={8}>
+        <Heading tag={titleTag} $fontSize={24} $lineHeight={"32px"} $mt={8}>
           <ActionLink href={href} title={title}>
             {title}
           </ActionLink>
         </Heading>
-        <P $fontSize={14} $mt={8}>
+        <P $fontSize={14} $mt={8} $mb={[16, 0]}>
           <LineClamp lines={2}>{snippet}</LineClamp>
         </P>
       </Flex>