+++ src/components/Forms/NewsletterForm/useNewsletterForm.ts
@@ -1,5 +1,3 @@
-import { useCallback } from "react";
-
 import { hubspotSubmitForm } from "../../../browser-lib/hubspot/forms";
 import { HubspotFormData } from "../../../browser-lib/hubspot/forms/hubspotSubmitForm";
 import config from "../../../config";
@@ -8,23 +6,28 @@ import useAnonymousId from "../../../browser-lib/analytics/useAnonymousId";
 
 const hubspotNewsletterFormId = config.get("hubspotNewsletterFormId");
 
-const useNewsletterForm = () => {
+type UseNewsletterFormProps = {
+  onSubmit?: () => void;
+};
+const useNewsletterForm = (props: UseNewsletterFormProps = {}) => {
   const anonymousId = useAnonymousId();
-
   const { identify } = useAnalytics();
-  const onSubmit = useCallback(
-    (data: HubspotFormData) => {
-      const hubspotFormResponse = hubspotSubmitForm({
-        hubspotFormId: hubspotNewsletterFormId,
-        data,
-      });
 
-      identify(anonymousId, { email: data.email || data.emailTextOnly });
+  const onSubmit = (data: HubspotFormData) => {
+    if (props.onSubmit) {
+      props.onSubmit();
+    }
 
-      return hubspotFormResponse;
-    },
-    [identify, anonymousId]
-  );
+    const hubspotFormResponse = hubspotSubmitForm({
+      hubspotFormId: hubspotNewsletterFormId,
+      data,
+    });
+
+    identify(anonymousId, { email: data.email || data.emailTextOnly });
+
+    return hubspotFormResponse;
+  };
+
   return { onSubmit };
 };