+++ src/node-lib/cms/sanity-client/schemas/pages.ts
@@ -1,14 +1,10 @@
 import * as z from "zod";
 
-import {
-  cardSchema,
-  CTASchema,
-  documentSchema,
-  imageSchema,
-  portableTextSchema,
-  textAndMediaSchema,
-  textBlockSchema,
-} from "./base";
+import { attachmentSchema, documentSchema, imageSchema } from "./base";
+import { cardSchema, textAndMediaSchema, textBlockSchema } from "./blocks";
+import { CTASchema } from "./cta";
+import { portableTextSchema } from "./portableText";
+import { teamMemberSchema } from "./teamMember";
 
 export const planningPageSchema = z
   .object({
@@ -43,17 +39,6 @@ export const planningPageSchema = z
 
 export type PlanningPage = z.infer<typeof planningPageSchema>;
 
-const attachmentSchema = z.object({
-  title: z.string(),
-  file: z.object({
-    asset: z.object({
-      extension: z.string(),
-      size: z.number(),
-      url: z.string(),
-    }),
-  }),
-});
-
 export const aboutPageSchema = z
   .object({
     title: z.string(),
@@ -77,6 +62,7 @@ export const aboutPageSchema = z
       introPortableText: portableTextSchema,
       documents: z.array(attachmentSchema),
       governancePortableText: portableTextSchema,
+      boardMembers: z.array(teamMemberSchema),
     }),
     partners: z.object({
       sectionHeading: z.string(),