+++ src/styles/reset.styles.ts
@@ -65,6 +65,10 @@ const resetStyles = css`
     font: inherit;
   }
 
+  cite {
+    font-style: inherit;
+  }
+
   /* Remove all animations, transitions and smooth scroll for people that prefer not to see them */
   @media (prefers-reduced-motion: reduce) {
     html:focus-within {